requirements.txt
openai
python-dotenv
langchain
langchain-openai
streamlit